Document-processing-function/requirements.txt
google-cloud-documentai==0.2.0
google-cloud-firestore==1.8.1
google-cloud-pubsub==1.7.0
google-cloud-storage==1.29.0